requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
git+https://github.com/jhj0517/jhj0517-whisper.git
faster-whisper
transformers
gradio==3.37.0
pytube
pyannote.audio
pydub